Add ability to get main sections and submission names from tidier data
Tidier data has to be addressed differently. So if the user runs `make_tidier_table()` and then tries to get the main sections or submission names, then the function would not return the correct values. Add parameter to acknowledge that the data is tidier for both functions. Add a separate, internal `get_main_sections_tidier()` function that is called if the data is tidier; this leaves the original function cleaner. Add testing.